+++ src/PluginClient.php
@@ -108,6 +108,10 @@ final class PluginClient implements HttpClient, HttpAsyncClient
      */
     private function configure(array $options = [])
     {
+        if (isset($options['debug_plugins'])) {
+            @trigger_error('The "debug_plugins" option is deprecated since 1.5 and will be removed in 2.0.', E_USER_DEPRECATED);
+        }
+
         $resolver = new OptionsResolver();
         $resolver->setDefaults([
             'max_restarts' => 10,