+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - `QueryDefaultsPlugin` to add default query parameters. 
 
+## Unreleased
+
+### Deprecated
+
+- The `debug_plugins` option for `PluginClient` is deprecated and will be removed in 2.0. Use the decorator design pattern instead like in [ProfilePlugin](https://github.com/php-http/HttplugBundle/blob/de33f9c14252f22093a5ec7d84f17535ab31a384/Collector/ProfilePlugin.php).
+
 ## 1.4.2 - 2017-03-18
 
 ### Deprecated