+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `PluginClientFactory` to create `PluginClient` instances.
 - Added new option 'delay' for `RetryPlugin`.
 - Added new option 'decider' for `RetryPlugin`.