+++ CHANGELOG.md
@@ -56,6 +56,10 @@ $plugin = new RetryPlugin(['delay' => function(RequestInterface $request, Except
 
 ## 1.3.0 - 2016-10-16
 
+### Added
+
+- Add HttpClientPool client to leverage load balancing and fallback mechanism [see the documentation](http://docs.php-http.org/en/latest/components/client-common.html) for more details
+
 ### Changed
 
 - Fix Emulated Trait to use Http based promise which respect the HttpAsyncClient interface